github: expect docker tag name to contain docker/
To distinguish between tags we push for a binary release (e.g. a new
version of lndinit itself) and a tag that we push just for bundling
lndinit with a newly released lnd version, we expect the docker tag to
start with docker/, and consequently only build images for tags with
that prefix.